requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
flask
torch==1.8.1+cpu 
torchvision==0.9.1+cpu
pathlib
pillow
gunicorn
numpy